[feat]Add node runtime support and tests.
