+++ .gitignore
@@ -1,1 +1,3 @@
 bower_components
+node_modules
+test/browser/runner.browser.js
