+++ .travis.yml
@@ -1,10 +1,21 @@
-script: "echo 1"
-after_success: |
-        [ $TRAVIS_BRANCH == 'master' ] && \
-        mkdir target && \
-        cp -r demo target/ && \
-        cp -r src target/ && \
-        cp -r plugin target/ && \
-        sudo pip install ghp-import && \
-        ghp-import -n target && \
-        git push -qf https://${TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git gh-pages
+language: node_js
+node_js:
+  - "0.12"
+sudo: false
+before_install:
+  - npm install bower grunt-cli -g
+  - bower install
+after_success:
+  - if [[ "$TRAVIS_BRANCH" == "master" ]] && [[ "${TRAVIS_PULL_REQUEST}" = "false" ]]; then
+      echo "Deploy demos to github pages.";
+      mkdir target;
+      cp -r demo target/;
+      cp -r dist target/;
+      cp -r plugin target/;
+      cd target && git init;
+      git config user.name "leancloud";
+      git config user.email "ci@leancloud.rocks";
+      git add .;
+      git commit -m "Deploy demos to Github Pages [skip ci]";
+      git push -qf https://${TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git master:gh-pages;
+    fi
