+++ demo/demo1/index.html
@@ -19,7 +19,7 @@
             <a target="_blank" href="https://leancloud.cn/docs/js_realtime.html">https://leancloud.cn/docs/js_realtime.html</a>
         </p>
     </div>
-    <script src="../../src/AV.realtime.js"></script>
+    <script src="../../dist/AV.realtime.js"></script>
     <script src="test.js"></script>
 </body>
 </html>
