+++ demo/demo2/index.html
@@ -43,7 +43,7 @@
     <!-- 引入插件部分结束 -->
 
     <!-- 引入 LeanCloud 实时通信 SDK -->
-    <script src="../../src/AV.realtime.js"></script>
+    <script src="../../dist/AV.realtime.js"></script>
     <!-- Demo 逻辑文件 -->
     <script src="test.js"></script>
 </body>
