+++ package.json
@@ -1,0 +1,43 @@
+{
+  "name": "js-realtime-sdk",
+  "version": "2.2.0",
+  "description": "LeanCloud Realtime Message JavaScript SDK",
+  "main": "src/realtime.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "test": "grunt test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/leeyeh/js-realtime-sdk.git"
+  },
+  "keywords": [
+    "leancloud",
+    "realtime",
+    "message"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/leeyeh/js-realtime-sdk/issues"
+  },
+  "homepage": "https://github.com/leeyeh/js-realtime-sdk#readme",
+  "devDependencies": {
+    "es6-promise": "^2.3.0",
+    "grunt": "^0.4.5",
+    "grunt-browserify": "^3.8.0",
+    "grunt-contrib-connect": "^0.10.1",
+    "grunt-mocha-phantomjs": "^0.7.0",
+    "grunt-simple-mocha": "^0.4.0",
+    "load-grunt-tasks": "^3.2.0",
+    "should": "^7.0.2",
+    "should-sinon": "0.0.3",
+    "sinon": "^1.15.4",
+    "ws": "^0.7.2"
+  },
+  "dependencies": {
+    "extend": "^3.0.0",
+    "superagent": "^1.2.0"
+  }
+}
