+++ test/browser/index.html
@@ -1,0 +1,31 @@
+<!doctype html>
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <title>Specs - LeanCloud Realtime JavaScript SDK</title>
+    <meta name="viewport" content="width=device-width">
+    <link rel="stylesheet" href="../../bower_components/mocha/mocha.css">
+</head>
+
+<body>
+    <div id="mocha"></div>
+    <script src="../../bower_components/mocha/mocha.js"></script>
+
+    <script>
+        mocha.setup({
+            ui: 'bdd',
+            timeout: 10000
+        });
+    </script>
+    <script src="runner.browser.js"></script>
+    <script>
+        if (window.mochaPhantomJS) {
+            mochaPhantomJS.run();
+        } else {
+            mocha.run();
+        }
+    </script>
+</body>
+
+</html>
