+++ test/browser/runner.js
@@ -1,0 +1,2 @@
+var realtime = require('../../src/realtime');
+require('../specs')(realtime);
