--- test/index.html
@@ -1,28 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <title>Specs - LeanCloud Realtime JavaScript SDK</title>
-    <meta name="viewport" content="width=device-width">
-    <link rel="stylesheet" href="../bower_components/mocha/mocha.css">
-</head>
-
-<body>
-    <div id="mocha"></div>
-    <script src="../bower_components/should/should.js"></script>
-    <script src="../bower_components/mocha/mocha.js"></script>
-    <script src="../src/AV.realtime.js"></script>
-
-    <script>
-        mocha.setup('bdd');
-    </script>
-    <script src="test.js"></script>
-    <script>
-        mocha.checkLeaks();
-        mocha.globals(['AV']);
-        mocha.run();
-    </script>
-</body>
-
-</html>
