+++ test/runner.js
@@ -1,0 +1,5 @@
+var realtime = require('../src/realtime');
+realtime.config({
+    WebSocket: require('ws')
+});
+require('./specs')(realtime);
